requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
plotly==5.11.0
scikit_learn==1.0.2
scipy==1.7.3
streamlit==1.15.2
